qduoj-choice-question-plugin/backend/requirements.txt
# 青岛OJ选择题插件依赖

# 树形结构支持
django-mptt>=0.13.4

# 标签系统支持
django-taggit>=2.1.0

# 数据导入导出
django-import-export>=3.0.0

# Excel文件处理
openpyxl>=3.0.9
xlrd>=2.0.1
xlwt>=1.3.0

# 数据验证
cerberus>=1.3.4

# 缓存支持
django-redis>=5.2.0

# 分页增强
django-filter>=22.1

# API文档生成
drf-spectacular>=0.25.1

# 权限管理
django-guardian>=2.4.0

# 统计分析
numpy>=1.21.0
pandas>=1.3.0

# 图片处理（如果需要题目包含图片）
Pillow>=9.0.0

# 文件处理
python-magic>=0.4.24